requirements.txt
aiogram==2.23.1
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
flake8==6.0.0
frozenlist==1.3.3
gTTS==2.3.0
idna==3.4
magic-filter==1.0.9
mccabe==0.7.0
multidict==6.0.4
pycodestyle==2.10.0
pydub==0.25.1
pyflakes==3.0.1
python-dotenv==0.21.0
pytz==2022.7
requests==2.28.1
six==1.16.0
SpeechRecognition==3.9.0
urllib3==1.26.13
yarl==1.8.2
